requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
elasticsearch==7.11.0
et-xmlfile==1.0.1
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
openpyxl==3.0.6
pandas==1.2.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
googlemaps
python-dotenv